sites: split control listener and management listener endpoints into separate files
